Rename UserPreferences -> UserSettings (#1018)
